build.zig: Fix thisDir for relative @src().file
While normal `zig build` invocations will continue to have absolute
paths for `@src.file`, when invoked by ZLS on my own system (NixOS
x86_64) it actually returns a relative path, breaking this project's
assumption that `thisDir` is absolute.

This commit fixes the issue by calculating a `projectPath` at runtime,
allocating memory and/or resolving paths if the result of
`@src().file` is non-absolute. This will require dependent projects to
invoke `flecs.pkg(b)` at runtime instead of simply adding a
comptime-known `flecs.pkg`, but it is a necessary evil to allow the
project to compile with a relative `@src().file`.