Fix component registration and ID handle creation
ID handles returned by `ecs_id_handle` are now consistent even when
the function is inlined or invoked at `comptime`. This is achieved by
extracting the container struct with the global `handle` variable to a
generic type `IdHandle(T)`.

Component registration no longer attempts to short circuit if the
component or tag being registered already has an ID stored in the
`ecs_id_handle` for that type. This behavior was incorrect as a
component can be shared across multiple worlds and needs to be
registered with the same ID in all of them.

The sentinel value for an unused component value is now 0 instead of
`maxInt(EcsId)`, as `EcsEntityDesc` interprets an `id` of 0 as a
request to generate a new entity ID. Additionally, to match the
behavior of `ECS_COMPONENT`, the `use_low_id` field is now set for
non-tag components to allow for some optimizations.